Ensure editing variable marks stage as changed
