+++ src/components/sections/Form/withFieldChangeHandlers.js
@@ -1,11 +1,12 @@
 import { connect } from 'react-redux';
 import { compose, withHandlers } from 'recompose';
+import { change } from 'redux-form';
 import { actionCreators as codebookActions } from '../../../ducks/modules/protocol/codebook';
 import { getTypeForComponent } from '../../Form/inputOptions';
 import { getCodebookProperties } from './helpers';
 
 const fieldChangeHandlers = withHandlers({
-  handleChangeFields: ({ createVariable, updateVariable, type, entity }) =>
+  handleChangeFields: ({ createVariable, updateVariable, type, entity, changeForm, form }) =>
     (values) => {
       const { variable, component, ...rest } = values;
       const variableType = getTypeForComponent(component);
@@ -15,6 +16,8 @@ const fieldChangeHandlers = withHandlers({
         type: variableType,
         ...codebookProperties,
       };
+
+      changeForm(form, '_modified', new Date().getTime()); // Register a change in the stage editor
 
       if (variable) {
         updateVariable(entity, type, variable, configuration);
@@ -27,6 +30,7 @@ const fieldChangeHandlers = withHandlers({
 });
 
 const mapDispatchToProps = {
+  changeForm: change,
   createVariable: codebookActions.createVariable,
   updateVariable: codebookActions.updateVariable,
 };
